backend/requirements.txt
# Screen capture
mss==9.0.2

# OCR
pytesseract==0.3.13
Pillow==10.4.0

# AI
anthropic==0.40.0

# API server
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Data models
pydantic==2.10.3
pydantic-settings==2.7.0

# HTTP client (scraper)
httpx==0.28.1

# Environment
python-dotenv==1.0.1

# Database
# SQLite is stdlib — no extra package needed

# Dev / test
pytest==8.3.4
pytest-asyncio==0.24.0
black==24.10.0
ruff==0.8.4
